requirements.txt
chris_plugin~=0.0.10
pybicpl~=0.3.0
numpy~=1.22.2
nibabel~=3.2.2
loguru~=0.6.0
